refacto(moderations): move repository internal code
