+++ packages/~/moderations/repository/src/get_duplicate_moderations.ts
@@ -1,0 +1,28 @@
+//
+
+import { schema, type MonComptePro_PgDatabase } from "@~/moncomptepro.database";
+import { and, asc, eq } from "drizzle-orm";
+
+//
+
+export async function get_duplicate_moderations(
+  pg: MonComptePro_PgDatabase,
+  {
+    organization_id,
+    user_id,
+  }: {
+    organization_id: number;
+    user_id: number;
+  },
+) {
+  return pg
+    .select()
+    .from(schema.moderations)
+    .where(
+      and(
+        eq(schema.moderations.organization_id, organization_id),
+        eq(schema.moderations.user_id, user_id),
+      ),
+    )
+    .orderBy(asc(schema.moderations.created_at));
+}
