+++ packages/~/moderations/repository/src/get_moderations_list.ts
@@ -1,6 +1,6 @@
 //
 
-import { schema } from "@~/moncomptepro.database";
+import { schema, type MonComptePro_PgDatabase } from "@~/moncomptepro.database";
 import {
   and,
   asc,
@@ -10,12 +10,11 @@ import {
   isNull,
   not,
 } from "drizzle-orm";
-import type { NodePgDatabase } from "drizzle-orm/node-postgres";
 
 //
 
 export function get_moderations_list(
-  pg: NodePgDatabase<typeof schema>,
+  pg: MonComptePro_PgDatabase,
   {
     search,
     pagination = { page: 0, take: 10 },
@@ -75,26 +74,4 @@ export function get_moderations_list(
       .where(where);
     return { moderations, count };
   });
-}
-
-export async function get_duplicate_moderations(
-  pg: NodePgDatabase<typeof schema>,
-  {
-    organization_id,
-    user_id,
-  }: {
-    organization_id: number;
-    user_id: number;
-  },
-) {
-  return pg
-    .select()
-    .from(schema.moderations)
-    .where(
-      and(
-        eq(schema.moderations.organization_id, organization_id),
-        eq(schema.moderations.user_id, user_id),
-      ),
-    )
-    .orderBy(asc(schema.moderations.created_at));
 }
