+++ packages/~/moderations/repository/tsconfig.json
@@ -1,0 +1,7 @@
+{
+  "compilerOptions": {
+    "outDir": "./node_modules/.cache/tsc"
+  },
+  "extends": "@~/config.typescript/base.json",
+  "references": [{ "path": "../../infra/moncomptepro/database/tsconfig.json" }]
+}
