+++ tsconfig.json
@@ -38,6 +38,7 @@
     { "path": "packages/~/infra/moncomptepro/database/tsconfig.json" },
     { "path": "packages/~/infra/zammad/api/tsconfig.json" },
     { "path": "packages/~/infra/zammad/lib/tsconfig.json" },
+    { "path": "packages/~/moderations/repository/tsconfig.json" },
     { "path": "packages/~/organizations/api/tsconfig.json" },
     { "path": "packages/~/users/api/tsconfig.json" }
   ]
